Completely Implemented Pure Pursuit
With a few catches of course:

I ported the implementation from ftclib, and they only allow odometry,  I modified it to use the PID but the current robot position does not get updated as the robot moves due to the functions being used.

Second the Drive.getPose() method does not return a rotation in the Pose 2d, which would mess up the pure pursuit algorithm.